fixing id Generation

+++ common/src/main/java/org/powertac/common/IdGenerator.java
@@ -35,7 +35,8 @@ public class IdGenerator
    */
   public static long createId() 
   {
-    return multiplier * prefix + counter++;
+    //return multiplier * prefix + counter++;
+    return (long)multiplier * (long)prefix + (long)counter++;
   }
 
   /**
